Move all files related to PlantUML and NLP to backend directory
