+++ backend/nlp/plantuml/MANIFEST.in
